feat: Controller now use trait for better use in EasyAdmin when used with other extensions
